refactor sort_values_table_py_entry
