+++ bodo/ir/sort.py
@@ -24,7 +24,6 @@ from bodo.libs.array import (
     cpp_table_to_py_data,
     delete_table,
     py_data_to_cpp_table,
-    sort_values_table_py_entry,
 )
 from bodo.transforms import distributed_analysis, distributed_pass
 from bodo.transforms.distributed_analysis import Distribution
@@ -511,6 +510,63 @@ def sort_table_for_interval_join(
             table_t,
             bounds_arr_t,
             types.boolean,
+            types.boolean,
+        ),
+        codegen,
+    )
+
+
+@intrinsic
+def sort_values_table_py_entry(
+    typingctx,
+    table_t,
+    n_keys_t,
+    vect_ascending_t,
+    na_position_b_t,
+    dead_keys_t,
+    n_rows_t,
+    bounds_t,
+    parallel_t,
+):
+    """
+    Interface to the sorting of tables.
+    """
+    from llvmlite import ir as lir
+
+    from bodo.libs.array import table_type
+
+    assert table_t == table_type, "C++ table type expected"
+
+    def codegen(context, builder, sig, args):
+        fnty = lir.FunctionType(
+            lir.IntType(8).as_pointer(),
+            [
+                lir.IntType(8).as_pointer(),
+                lir.IntType(64),
+                lir.IntType(8).as_pointer(),
+                lir.IntType(8).as_pointer(),
+                lir.IntType(8).as_pointer(),
+                lir.IntType(8).as_pointer(),
+                lir.IntType(8).as_pointer(),
+                lir.IntType(1),
+            ],
+        )
+        fn_tp = cgutils.get_or_insert_function(
+            builder.module, fnty, name="sort_values_table_py_entry"
+        )
+        ret = builder.call(fn_tp, args)
+        bodo.utils.utils.inlined_check_and_propagate_cpp_exception(context, builder)
+        return ret
+
+    return (
+        table_type(
+            table_t,
+            types.int64,
+            types.voidptr,
+            types.voidptr,
+            types.voidptr,
+            types.voidptr,
+            types.voidptr,
             types.boolean,
         ),
         codegen,
