+++ bodo/libs/array.py
@@ -3039,59 +3039,6 @@ def interval_join_table(
 
 
 @intrinsic
-def sort_values_table_py_entry(
-    typingctx,
-    table_t,
-    n_keys_t,
-    vect_ascending_t,
-    na_position_b_t,
-    dead_keys_t,
-    n_rows_t,
-    bounds_t,
-    parallel_t,
-):
-    """
-    Interface to the sorting of tables.
-    """
-    assert table_t == table_type, "C++ table type expected"
-
-    def codegen(context, builder, sig, args):
-        fnty = lir.FunctionType(
-            lir.IntType(8).as_pointer(),
-            [
-                lir.IntType(8).as_pointer(),
-                lir.IntType(64),
-                lir.IntType(8).as_pointer(),
-                lir.IntType(8).as_pointer(),
-                lir.IntType(8).as_pointer(),
-                lir.IntType(8).as_pointer(),
-                lir.IntType(8).as_pointer(),
-                lir.IntType(1),
-            ],
-        )
-        fn_tp = cgutils.get_or_insert_function(
-            builder.module, fnty, name="sort_values_table_py_entry"
-        )
-        ret = builder.call(fn_tp, args)
-        bodo.utils.utils.inlined_check_and_propagate_cpp_exception(context, builder)
-        return ret
-
-    return (
-        table_type(
-            table_t,
-            types.int64,
-            types.voidptr,
-            types.voidptr,
-            types.voidptr,
-            types.voidptr,
-            types.voidptr,
-            types.boolean,
-        ),
-        codegen,
-    )
-
-
-@intrinsic
 def shuffle_renormalization(typingctx, table_t, random_t, random_seed_t, is_parallel_t):
     """
     Interface to the rebalancing of the table
